requirements.txt
#
####### requirements.txt #######
#
###### Requirements without Version Specifiers ######
requests
beautifulsoup4
firebase-admin
python-telegram-bot
google-cloud
google-cloud-vision